Move camera session out of CamConfig
