+++ app/src/main/java/app/grapheneos/camera/CamConfig.kt
@@ -2,40 +2,24 @@ package app.grapheneos.camera
 
 import android.annotation.SuppressLint
 import android.net.Uri
-import android.os.Handler
-import android.os.Looper
 import android.provider.MediaStore
 import android.util.Log
-import android.util.Size
 import androidx.annotation.VisibleForTesting
 import androidx.camera.core.AspectRatio
 import androidx.camera.core.Camera
-import androidx.camera.core.CameraInfo
 import androidx.camera.core.CameraSelector
-import androidx.camera.core.DynamicRange
 import androidx.camera.core.ImageAnalysis
 import androidx.camera.core.ImageCapture
 import androidx.camera.core.Preview
-import androidx.camera.core.SessionConfig
-import androidx.camera.core.TorchState
-import androidx.camera.core.UseCase
 import androidx.camera.core.ZoomState
 import androidx.camera.core.featuregroup.GroupableFeature
-import androidx.camera.core.resolutionselector.ResolutionSelector
-import androidx.camera.core.resolutionselector.ResolutionStrategy
-import androidx.camera.extensions.ExtensionMode
-import androidx.camera.extensions.ExtensionsManager
 import androidx.camera.lifecycle.ProcessCameraProvider
 import androidx.camera.video.Quality
 import androidx.camera.video.Recorder
 import androidx.camera.video.VideoCapture
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.Observer
-import app.grapheneos.camera.analyzer.QRAnalyzer
-import app.grapheneos.camera.data.camera.model.ExtensionKey
-import app.grapheneos.camera.data.camera.repository.CameraProviderSource
-import app.grapheneos.camera.data.camera.repository.FeatureCombinationSupport
-import app.grapheneos.camera.data.camera.store.ExtensionAvailabilityStore
+import app.grapheneos.camera.data.camera.model.CameraBindSettings
+import app.grapheneos.camera.data.camera.repository.CameraSessionEnvironment
+import app.grapheneos.camera.data.camera.repository.CameraSession
 import app.grapheneos.camera.data.core.model.CameraMode
 import app.grapheneos.camera.data.media.repository.CapturedItemRepository
 import app.grapheneos.camera.data.settings.model.CameraSettings
@@ -43,24 +27,16 @@ import app.grapheneos.camera.data.settings.model.GridType
 import app.grapheneos.camera.data.settings.model.ModeSettings
 import app.grapheneos.camera.data.settings.model.SettingsDefaults
 import app.grapheneos.camera.ui.videoQualityTitle
+import app.grapheneos.camera.ui.viewfinder.ViewfinderEffects
 import app.grapheneos.camera.data.settings.repository.SettingsRepository
-import app.grapheneos.camera.domain.camera.mapper.VideoQualityFeatureMapper
-import app.grapheneos.camera.domain.camera.model.CameraBindRequest
 import app.grapheneos.camera.domain.camera.model.CameraEntryPoint
-import app.grapheneos.camera.domain.camera.model.FeatureGroupRequest
-import app.grapheneos.camera.domain.camera.model.ImageCaptureMode
-import app.grapheneos.camera.domain.camera.model.InVideoSnapshotSupport
-import app.grapheneos.camera.domain.camera.usecase.BuildCameraSessionPlan
 import app.grapheneos.camera.domain.camera.usecase.ResolveAvailableModes
 import app.grapheneos.camera.domain.camera.usecase.ResolveDroppedVideoQuality
-import app.grapheneos.camera.domain.camera.usecase.ResolveInVideoSnapshotSupport
 import com.google.zxing.BarcodeFormat
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 import java.io.IOException
-import java.util.concurrent.Executors
-import kotlin.concurrent.thread
 import kotlin.properties.ReadWriteProperty
 import kotlin.reflect.KProperty
 import kotlinx.coroutines.CoroutineScope
@@ -73,18 +49,13 @@ import kotlinx.coroutines.runBlocking
 @SuppressLint("UnsafeOptInUsageError")
 class CamConfig @AssistedInject constructor(
     @Assisted private val environment: CameraSessionEnvironment,
-    @Assisted private val effects: CameraSessionEffects,
+    @Assisted private val effects: ViewfinderEffects,
     private val entryPoint: CameraEntryPoint,
     private val settingsRepository: SettingsRepository,
     private val capturedItemRepository: CapturedItemRepository,
-    private val cameraProviderSource: CameraProviderSource,
-    private val extensionAvailabilityStore: ExtensionAvailabilityStore,
-    private val featureCombinationSupport: FeatureCombinationSupport,
-    private val buildCameraSessionPlan: BuildCameraSessionPlan,
-    private val videoQualityFeatureMapper: VideoQualityFeatureMapper,
-    private val resolveInVideoSnapshotSupport: ResolveInVideoSnapshotSupport,
     private val resolveAvailableModes: ResolveAvailableModes,
     private val resolveDroppedVideoQuality: ResolveDroppedVideoQuality,
+    cameraSessionFactory: CameraSession.Factory,
 ) {
 
     @AssistedFactory
@@ -92,7 +63,7 @@ class CamConfig @AssistedInject constructor(
 
         fun create(
             environment: CameraSessionEnvironment,
-            effects: CameraSessionEffects,
+            effects: ViewfinderEffects,
         ): CamConfig
     }
 
@@ -118,95 +89,101 @@ class CamConfig @AssistedInject constructor(
 
         val DEFAULT_CAMERA_MODE = CameraMode.CAMERA
 
-        val FRONT_CAMERA_SELECTOR = CameraSelector.Builder()
-            .requireLensFacing(CameraSelector.LENS_FACING_FRONT)
-            .build()
-
-        val REAR_CAMERA_SELECTOR = CameraSelector.Builder()
-            .requireLensFacing(CameraSelector.LENS_FACING_BACK)
-            .build()
-
-        // Every setting that reaches one of the three probed SessionConfigs has to appear in the
-        // key. A setting added to the ImageCapture, Recorder or Preview builder without being added
-        // here would be answered from a verdict that predates it, which either takes in-video
-        // snapshots away for no reason or keeps them on a camera that cannot bind them.
-        private val snapshotSupport = HashMap<SnapshotProbeKey, InVideoSnapshotSupport>()
-
-        // A cache hit and a repeated probe reach the same verdict, so this is the only thing that
-        // tells them apart from the outside.
         @VisibleForTesting
-        var snapshotProbeCount = 0
-            private set
+        val snapshotProbeCount: Int
+            get() {
+                return CameraSession.snapshotProbeCount
+            }
 
         @VisibleForTesting
         fun clearSnapshotProbeCache() {
-            snapshotSupport.clear()
-            snapshotProbeCount = 0
+            CameraSession.clearSnapshotProbeCache()
         }
-
-        // The provider the verdicts above were probed through. A different instance means the
-        // camera stack was reinitialized and none of them describe it any more.
-        private var probedCameraProvider: ProcessCameraProvider? = null
     }
 
-    private data class SnapshotProbeKey(
-        val lensFacing: Int,
-        val videoQuality: Quality,
-        val usesFeatureGroup: Boolean,
-        val captureMode: ImageCaptureMode,
-        val selectHighestResolution: Boolean,
+    private val session = cameraSessionFactory.create(
+        environment = environment,
+        listener = object : CameraSession.Listener {
+            override fun onZoomStateChanged() {
+                effects.updateZoomThumb()
+            }
+
+            override fun onCameraProviderUnavailable() {
+                effects.showMessage(R.string.camera_provider_init_failure)
+            }
+
+            override fun onExtensionsUnavailable() {
+                effects.showMessage(R.string.extensions_manager_init_failure)
+            }
+
+            override fun onProviderReady(forced: Boolean) {
+                startCamera(forced = forced)
+            }
+
+            override fun onFeaturesSelected(
+                boundLensFacing: Int,
+                requested: List<GroupableFeature>,
+                qualityFeature: GroupableFeature?,
+                selected: Set<GroupableFeature>,
+            ) {
+                this@CamConfig.onFeaturesSelected(
+                    boundLensFacing, requested, qualityFeature, selected,
+                )
+            }
+        },
     )
 
-    var camera: Camera? = null
-
-    // Asking CameraInfo for the zoom state is cheap for a plain camera but costs ~100 ms once an
-    // extension is bound, because CameraX then queries the extension's zoom range through
-    // CameraExtensionCharacteristics, which enumerates every vendor key. Read this snapshot instead
-    // of the camera on any path that runs more than once per bind. It is null from the moment a
-    // bind starts until attachZoomState has run, which is where that query was moved to.
-    var zoomState: ZoomState? = null
-        private set
-
-    private var zoomStateSource: LiveData<ZoomState>? = null
-
-    private val zoomStateObserver = Observer<ZoomState> {
-        zoomState = it
-        if (it.linearZoom != 0f || it.zoomRatio != 1f) {
-            effects.updateZoomThumb()
-        }
-    }
-
-    private val handler = Handler(Looper.getMainLooper())
-
-    private val attachZoomState = Runnable {
-        if (!environment.isSessionActive) return@Runnable
-
-        zoomStateSource = camera?.cameraInfo?.zoomState?.also {
-            it.observe(environment.sessionLifecycleOwner, zoomStateObserver)
+    var camera: Camera?
+        get() = session.camera
+        set(value) {
+            session.camera = value
         }
 
-        zoomState = zoomStateSource?.value
+    val zoomState: ZoomState?
+        get() = session.zoomState
+
+    val cameraProvider: ProcessCameraProvider?
+        get() = session.cameraProvider
+
+    val imageCapture: ImageCapture?
+        get() = session.imageCapture
+
+    val preview: Preview?
+        get() = session.preview
+
+    val videoCapture: VideoCapture<Recorder>?
+        get() = session.videoCapture
+
+    val iAnalyzer: ImageAnalysis?
+        get() = session.iAnalyzer
+
+    var lensFacing: Int
+        get() = session.lensFacing
+        set(value) {
+            session.lensFacing = value
+        }
+
+    val isFlashAvailable: Boolean
+        get() = session.isFlashAvailable
+
+    var isTorchOn: Boolean
+        get() = session.isTorchOn
+        set(value) {
+            session.isTorchOn = value
+        }
+
+    val isZslSupported: Boolean
+        get() = session.isZslSupported
+
+    fun canApplyVideoStabilization(): Boolean {
+        return session.canApplyVideoStabilization()
     }
 
-    var cameraProvider: ProcessCameraProvider? = null
-    private var extensionsManager: ExtensionsManager? = null
-
-    var imageCapture: ImageCapture? = null
-        private set
-
-    var preview: Preview? = null
+    fun toggleTorchState() {
+        session.toggleTorchState()
+    }
 
     val allowedFormats: ArrayList<BarcodeFormat> = arrayListOf()
-
-    private val cameraExecutor by lazy {
-        Executors.newSingleThreadExecutor()
-    }
-
-    var videoCapture: VideoCapture<Recorder>? = null
-
-    private var qrAnalyzer: QRAnalyzer? = null
-
-    var iAnalyzer: ImageAnalysis? = null
 
     @set:VisibleForTesting
     var mPlayer = environment.createTunePlayer()
@@ -292,22 +269,6 @@ class CamConfig @AssistedInject constructor(
     var isQRMode = false
         private set
 
-    val isFlashAvailable: Boolean
-        get() = camera?.cameraInfo?.hasFlashUnit() ?: false
-
-    var isTorchOn: Boolean = false
-        get() {
-            return camera?.cameraInfo?.torchState?.value == TorchState.ON
-        }
-        set(value) {
-            field = if (isFlashAvailable) {
-                camera?.cameraControl?.enableTorch(value)
-                value
-            } else {
-                false
-            }
-        }
-
     var currentMode: CameraMode = DEFAULT_CAMERA_MODE
         private set
 
@@ -332,12 +293,6 @@ class CamConfig @AssistedInject constructor(
                 settingsRepository.update { it.copy(aspectRatio = value) }
             }
         }
-
-    var lensFacing = DEFAULT_LENS_FACING
-
-    private var cameraSelector: CameraSelector = CameraSelector.Builder()
-        .requireLensFacing(DEFAULT_LENS_FACING)
-        .build()
 
     var gridType: GridType by setting(
         read = { it.gridType },
@@ -399,7 +354,7 @@ class CamConfig @AssistedInject constructor(
                 effects.applyScanAllCodesChrome(value)
             }
 
-            qrAnalyzer?.refreshHints()
+            session.refreshQrHints()
         }
 
     var includeAudio: Boolean
@@ -461,39 +416,6 @@ class CamConfig @AssistedInject constructor(
         read = { it.gyroscopeSuggestions },
         write = { current, value -> current.copy(gyroscopeSuggestions = value) },
     )
-
-    val isZslSupported: Boolean by lazy {
-        camera!!.cameraInfo.isZslSupported
-    }
-
-    // Whether the EIS toggle has anything to act on. Stabilization is only ever requested through
-    // the feature group (see startCamera), which in turn needs the platform to be able to verify
-    // feature combinations, so a camera that can stabilize is not on its own enough: where the
-    // group cannot be used nothing applies EIS, and offering the toggle there is offering a
-    // control that does nothing.
-    fun canApplyVideoStabilization(): Boolean {
-        return canVerifyFeatureCombinations() && isVideoStabilizationSupported()
-    }
-
-    private fun isVideoStabilizationSupported(): Boolean {
-        // The toggle asks for both kinds of stabilization (see the preferred feature group in
-        // startCamera), preferring the preview kind and falling back to the recording-only kind,
-        // so it is meaningful whenever either one is available. Testing only the recorder
-        // capability both hid the toggle on cameras that can stabilize the preview but not the
-        // recording, and offered it on cameras where only the recorder can stabilize - where
-        // the bind used to ask exclusively for preview stabilization, leaving the toggle with
-        // nothing to do.
-        return isPreviewStabilizationSupported() || isRecorderStabilizationSupported()
-    }
-
-    private fun isPreviewStabilizationSupported(): Boolean {
-        return Preview.getPreviewCapabilities(getCurrentCameraInfo()).isStabilizationSupported
-    }
-
-
-    private fun isRecorderStabilizationSupported(): Boolean {
-        return Recorder.getVideoCapabilities(getCurrentCameraInfo()).isStabilizationSupported
-    }
 
     fun shouldShowGyroscope(): Boolean {
         return isInPhotoMode && gSuggestions
@@ -570,7 +492,7 @@ class CamConfig @AssistedInject constructor(
             }
         }
 
-        qrAnalyzer?.refreshHints()
+        session.refreshQrHints()
     }
 
     private fun slotCurrentMode() {
@@ -613,7 +535,7 @@ class CamConfig @AssistedInject constructor(
 
         effects.selectBarcodeFormatToggles(allowedFormats)
 
-        qrAnalyzer?.refreshHints()
+        session.refreshQrHints()
     }
 
     var waitForFocusLock: Boolean by setting(
@@ -625,10 +547,6 @@ class CamConfig @AssistedInject constructor(
         read = { it.selectHighestResolution },
         write = { current, value -> current.copy(selectHighestResolution = value) },
     )
-
-    fun toggleTorchState() {
-        isTorchOn = !isTorchOn
-    }
 
     fun toggleFlashMode() {
         if (isFlashAvailable) {
@@ -655,10 +573,6 @@ class CamConfig @AssistedInject constructor(
         startCamera(true)
     }
 
-    private fun getCurrentCameraInfo(): CameraInfo {
-        return cameraProvider!!.getCameraInfo(cameraSelector)
-    }
-
     fun toggleCameraSelector() {
 
         // Manually switch to the opposite lens facing
@@ -671,7 +585,7 @@ class CamConfig @AssistedInject constructor(
 
         // Test whether the new lens facing is supported by the current device
         // If it is supported then restart the camera with the new configuration
-        if (isLensFacingSupported(lensFacing)) {
+        if (session.isLensFacingSupported(lensFacing, currentMode.extensionMode)) {
             startCamera(true)
         } else {
             // Else revert back to the old facing (while displaying an error message
@@ -687,209 +601,6 @@ class CamConfig @AssistedInject constructor(
 
     }
 
-    fun initializeCamera(forced: Boolean = false) {
-        if (cameraProvider != null) {
-            startCamera(forced = forced)
-            return
-        }
-
-        cameraProviderSource.acquireProvider(environment.sessionContext) { provider ->
-            when (provider) {
-                null -> effects.showMessage(R.string.camera_provider_init_failure)
-
-                else -> onCameraProviderReady(provider, forced)
-            }
-        }
-    }
-
-    private fun onCameraProviderReady(provider: ProcessCameraProvider, forced: Boolean) {
-        if (provider !== probedCameraProvider) {
-            // A different provider instance means the camera stack was reinitialized:
-            // extension verdicts probed through the previous instance (including bind-time
-            // blacklists, see startCamera) describe vendor state that no longer exists.
-            extensionAvailabilityStore.clear()
-            snapshotSupport.clear()
-            probedCameraProvider = provider
-        }
-        cameraProvider = provider
-
-        // Manually switch to the other lens facing (if the default lens facing isn't
-        // supported for the current device)
-        if (!isLensFacingSupported(lensFacing)) {
-            lensFacing = if (lensFacing == CameraSelector.LENS_FACING_BACK) {
-                CameraSelector.LENS_FACING_FRONT
-            } else {
-                CameraSelector.LENS_FACING_BACK
-            }
-        }
-
-        cameraProviderSource.acquireExtensionsManager(
-            environment.sessionContext,
-            provider,
-        ) { manager ->
-            if (manager == null) {
-                effects.showMessage(R.string.extensions_manager_init_failure)
-            } else {
-                extensionsManager = manager
-            }
-
-            startCamera(forced = forced)
-        }
-    }
-
-    // ExtensionsManager.isExtensionAvailable() answers from CameraExtensionCharacteristics'
-    // static advertisement data. Actually *binding* an advertised extension additionally makes
-    // CameraX initialize a Camera2ExtensionsVendorExtender, which calls into the vendor's
-    // advanced extender over binder. Some vendors advertise a mode there and then throw from
-    // that init - Pixels raise "Framework size list map not supported in pixel path"
-    // - which used to kill the process from inside bindToLifecycle(). CameraX 1.6 removed the
-    // legacy OEM extender path, so there is no longer a working fallback on those devices and the
-    // only safe option is to stop offering the mode.
-    //
-    // getCameraInfo() performs exactly the same vendor init that bindToLifecycle() does, so it is
-    // a faithful probe. Verdicts are cached (ExtensionAvailabilityStore) because every step of
-    // the probe -- including the availability query, see probeExtension -- costs a binder round
-    // trip. A negative verdict is only cached when the failure is known to be persistent: caching
-    // a transient probe failure would make the mode's tab vanish for the rest of the process
-    // lifetime over a condition that clears seconds later.
-    //
-    // The cache is read and written on the main thread only, which is also what keeps the two
-    // activities that can share it (a secure session over a running one) from racing each other.
-    // probeExtension() itself also runs on the probe thread that loadTabs() spawns, but its
-    // results come back through the main executor.
-    //
-    // true/false is a verdict that is safe to cache; null is a failure that may be transient
-    // and must not be (see ExtensionAvailabilityStore). The provider and manager are parameters
-    // rather than the fields because this also runs off the main thread, where the fields could
-    // be swapped out mid-probe.
-    private fun probeExtension(
-        provider: ProcessCameraProvider,
-        em: ExtensionsManager,
-        selector: CameraSelector,
-        extensionMode: Int,
-    ): Boolean? {
-        // What the vendor advertises is as static as the vendor init verdict below, so a
-        // negative answer is cached the same way -- notably, isExtensionAvailable() is *not*
-        // the cheap in-process lookup it appears to be: on Android 17 every call costs a round
-        // trip to the vendor's extensions proxy service, which is exactly the kind of work this
-        // probe exists to keep off the main thread (see loadTabs). It sits inside the try below so
-        // that a transient failure of that round trip is treated like any other -- returning null
-        // to retry later -- rather than propagating out: on the background probe thread an escaping
-        // exception would strand the round with extensionProbesInFlight still set, blocking every
-        // later tab refresh.
-        return try {
-            when {
-                !em.isExtensionAvailable(selector, extensionMode) -> false
-                else -> {
-                    provider.getCameraInfo(
-                        em.getExtensionEnabledCameraSelector(selector, extensionMode)
-                    )
-                    true
-                }
-            }
-        } catch (e: UnsupportedOperationException) {
-            // The signature of a vendor extender that advertises the mode and then throws from
-            // its own init (Pixels: "Framework size list map not supported in pixel path").
-            // Nothing about it changes within a process lifetime, so this verdict is safe to
-            // remember.
-            Log.w(TAG, "Extension mode $extensionMode is advertised but unusable here", e)
-            false
-        } catch (e: Exception) {
-            // Anything else may be transient — the camera service restarting, the camera briefly
-            // held by another process. Fail this probe but leave the cache alone so the mode is
-            // offered again once the underlying condition clears.
-            Log.w(TAG, "Probing extension mode $extensionMode failed, will retry later", e)
-            null
-        }
-    }
-
-    private fun isExtensionUsable(
-        selector: CameraSelector,
-        lensFacing: Int,
-        extensionMode: Int,
-        probeOnMiss: Boolean = true,
-    ): Boolean {
-        if (extensionMode == ExtensionMode.NONE) return true
-
-        val em = extensionsManager ?: return false
-        val provider = cameraProvider ?: return false
-
-        val key = ExtensionKey(
-            lensFacing = lensFacing,
-            extensionMode = extensionMode,
-        )
-        extensionAvailabilityStore.verdict(key)?.let { return it }
-
-        if (!probeOnMiss) return false
-
-        // A background probe round (loadTabs) may already be asking the vendor about this very
-        // key. Probing inline here would run a second synchronous vendor round trip on the main
-        // thread and race that round's verdict write. While a round is in flight, treat the miss
-        // as "unusable for now"; the round fills the cache and the next rebind/tab refresh picks
-        // up the real verdict. The inline probe stays for the not-in-flight cold case, where it is
-        // the only path to a verdict.
-        if (extensionProbesInFlight) return false
-
-        val verdict = probeExtension(provider, em, selector, extensionMode) ?: return false
-        extensionAvailabilityStore.record(key, usable = verdict)
-        return verdict
-    }
-
-    private fun canVerifyFeatureCombinations(): Boolean {
-        val cameraInfo = try {
-            cameraProvider?.getCameraInfo(cameraSelector)
-        } catch (exception: IllegalArgumentException) {
-            Log.w(TAG, "Unable to resolve camera info for feature combination gate", exception)
-            null
-        }
-
-        return when (cameraInfo) {
-            null -> false
-            else -> featureCombinationSupport.canVerify(cameraInfo)
-        }
-    }
-
-    // Maps the user-chosen video quality to the equivalent groupable feature, for use when a
-    // feature group is passed to SessionConfig (see startCamera). Quality.HIGHEST has no
-    // groupable equivalent and is resolved to the highest quality the current camera supports,
-    // mirroring what QualitySelector.from(Quality.HIGHEST) would have selected.
-    private fun videoQualityAsGroupableFeature(): GroupableFeature? {
-        val quality = when (videoQuality) {
-            Quality.HIGHEST -> highestSupportedVideoQuality()
-            else -> videoQuality
-        } ?: return null
-
-        val feature = videoQualityFeatureMapper.map(quality)
-
-        if (feature == null) {
-            // Not fatal: startCamera() then requests no quality feature at all and the quality is
-            // left to Recorder's default selector. Worth a log because it means the user's
-            // explicit choice is silently not being asked for.
-            Log.w(TAG, "No groupable feature equivalent for video quality $quality")
-        }
-
-        return feature
-    }
-
-    private fun highestSupportedVideoQuality(): Quality? {
-        val cameraInfo = try {
-            cameraProvider?.getCameraInfo(cameraSelector)
-        } catch (exception: IllegalArgumentException) {
-            Log.w(TAG, "Unable to resolve camera info for quality lookup", exception)
-            null
-        }
-
-        return cameraInfo?.let {
-            Recorder.getVideoCapabilities(it)
-                .getSupportedQualities(DynamicRange.SDR)
-                .firstOrNull()
-        }
-    }
-
-    // What was asked for -- including which camera it was asked of -- is passed in rather than
-    // read back from a field: the callback is delivered asynchronously, so a field could already
-    // describe a later bind by the time this runs, and the message would then name settings (or
-    // dedup against a camera) that this result never involved.
     private fun onFeaturesSelected(
         boundLensFacing: Int,
         requested: List<GroupableFeature>,
@@ -915,32 +626,11 @@ class CamConfig @AssistedInject constructor(
         )
     }
 
-    private fun isLensFacingSupported(lensFacing: Int): Boolean {
-        var tCameraSelector = CameraSelector.Builder()
-            .requireLensFacing(lensFacing)
-            .build()
-
-        if (currentMode.extensionMode != ExtensionMode.NONE) {
-            extensionsManager?.let { em ->
-                if (!isExtensionUsable(tCameraSelector, lensFacing, currentMode.extensionMode))
-                    return false
-
-                try {
-                    tCameraSelector = em.getExtensionEnabledCameraSelector(
-                        tCameraSelector,
-                        currentMode.extensionMode
-                    )
-                } catch (e: IllegalArgumentException) {
-                    return false
-                }
-            }
-        }
-
-        return cameraProvider?.hasCamera(tCameraSelector) ?: false
+    fun initializeCamera(forced: Boolean = false) {
+        session.initialize(forced = forced, extensionMode = currentMode.extensionMode)
     }
 
     // Start the camera with latest hard configuration
-    @SuppressLint("RestrictedApi")
     fun startCamera(forced: Boolean = false) {
         if ((!forced && camera != null) || cameraProvider == null) return
 
@@ -963,7 +653,7 @@ class CamConfig @AssistedInject constructor(
         // (Snackbar/popup message can be shown before startCamera is called
         // in specific cases of explicitly switching to another side or if
         // the camera is expected)
-        if (!isLensFacingSupported(lensFacing)) {
+        if (!session.isLensFacingSupported(lensFacing, currentMode.extensionMode)) {
             lensFacing = if (lensFacing == CameraSelector.LENS_FACING_BACK) {
                 CameraSelector.LENS_FACING_FRONT
             } else {
@@ -972,272 +662,82 @@ class CamConfig @AssistedInject constructor(
         }
 
 
-        cameraSelector = CameraSelector.Builder()
-            .requireLensFacing(lensFacing)
-            .build()
+        session.selectLensFacing(lensFacing)
 
         // To use the last frame instead of showing a blank screen when
         // the camera that is being currently used gets unbind
         effects.updateLastFrame()
 
-        // Unbind/close all other camera(s) [if any]
-        cameraProvider?.unbindAll()
-
-        val extMode = currentMode.extensionMode
-        var appliedExtension: ExtensionKey? = null
-        if (extMode != ExtensionMode.NONE) {
-            val em = extensionsManager
-            if (em != null && isExtensionUsable(cameraSelector, lensFacing, extMode)) {
-                appliedExtension = ExtensionKey(
-                    lensFacing = lensFacing,
-                    extensionMode = extMode,
-                )
-                cameraSelector = em.getExtensionEnabledCameraSelector(cameraSelector, extMode)
-            } else {
-                Log.e(TAG, "Mode $currentMode isn't available for this device")
+        val qrLensFacing = when {
+            isQRMode -> {
+                effects.startFocusTimer()
+                qrLensFacing()
             }
+
+            else -> null
         }
 
-        val useCasesList = arrayListOf<UseCase>()
-
-        // CameraX 1.6.0's SessionConfig throws an IllegalArgumentException at construction time
-        // if any use case configures a groupable feature through a non-groupable API while a
-        // feature group is in use. Recorder.Builder.setQualitySelector() is such an API since
-        // 1.6.0 introduced GroupableFeatures.*_RECORDING, so when EIS is requested through
-        // GroupableFeature.PREVIEW_STABILIZATION (the only case where this app uses a feature
-        // group), the video quality has to be requested through the feature group as well.
-        //
-        // The validation triggers on setQualitySelector() having been called at all, not on the
-        // quality it was given, so this flag -- not the resolved feature below -- is what decides
-        // whether that setter may be used. videoQualityAsGroupableFeature() can legitimately fail
-        // to map the current quality, and falling back to setQualitySelector() in that case would
-        // reintroduce the very exception this works around.
-        //
-        // Both halves of canApplyVideoStabilization() gate the group. The platform has to be able
-        // to verify feature combinations: where it cannot, the resolver would conflate "could not
-        // check" with "unsupported", quietly discard the stored video quality and produce
-        // untruthful notices. Cameras behind that gate use the pre-1.6 non-groupable setters in
-        // the session plan instead, which are legal exactly because no feature group is in use
-        // then. And the
-        // group exists solely to negotiate stabilization against the quality, so a camera that
-        // supports no stabilization at all has nothing to negotiate: it takes the plain fallback
-        // path directly, which produces the identical output without a needless feature-group
-        // round.
-        val featureGroup: FeatureGroupRequest = when {
-            isVideoMode && enableEIS && canApplyVideoStabilization() -> {
-                FeatureGroupRequest.Requested(videoQualityAsGroupableFeature())
-            }
-
-            else -> FeatureGroupRequest.Unused
-        }
-
-        val requiredQualityFeature = when (featureGroup) {
-            is FeatureGroupRequest.Requested -> featureGroup.videoQualityFeature
-            FeatureGroupRequest.Unused -> null
-        }
-
-        val bindRequest = CameraBindRequest(
-            includesVideoCapture = !isQRMode && isVideoMode,
-            includesImageCapture = !isQRMode && !entryPoint.requiresVideoModeOnly,
-            aspectRatio = aspectRatio,
-            imageCaptureTargetRotation = imageCapture?.targetRotation ?: rotation,
-            previewTargetRotation = preview?.targetRotation ?: rotation,
-            flashMode = flashMode,
-            photoQuality = photoQuality,
-            waitForFocusLock = waitForFocusLock,
-            enableZsl = enableZsl,
-            selectHighestResolution = selectHighestResolution,
-            videoQuality = videoQuality,
-            mirrorVideoOnFrontCamera = saveVideoAsPreviewed,
-            featureGroup = featureGroup,
-        )
-
-        val plan = buildCameraSessionPlan(bindRequest)
-
-        if (isQRMode) {
-            val analyzer = environment.createQrAnalyzer()
-            val strategy = ResolutionStrategy(
-                Size(960, 960),
-                ResolutionStrategy.FALLBACK_RULE_CLOSEST_HIGHER_THEN_LOWER
-            )
-            val mIAnalyzer = ImageAnalysis.Builder()
-                .setResolutionSelector(
-                    ResolutionSelector.Builder().setResolutionStrategy(strategy).build()
-                )
-                .setOutputImageRotationEnabled(true)
-                .build()
-            qrAnalyzer = analyzer
-            effects.startFocusTimer()
-            iAnalyzer = mIAnalyzer
-            mIAnalyzer.setAnalyzer(cameraExecutor, analyzer)
-            cameraSelector = CameraSelector.Builder()
-                .requireLensFacing(
-                    if (isLensFacingSupported(CameraSelector.LENS_FACING_BACK)) {
-                        CameraSelector.LENS_FACING_BACK
-                    } else {
-                        effects.showMessage(R.string.qr_rear_camera_unavailable)
-                        CameraSelector.LENS_FACING_FRONT
-                    }
-                )
-                .build()
-            useCasesList.add(mIAnalyzer)
-
-        } else {
-            if (isVideoMode) {
-                effects.setMicMutedIconVisible(!includeAudio)
-            }
-
-            plan.videoCapture?.let {
-                videoCapture = it
-                useCasesList.add(it)
-            }
-
-            plan.imageCapture?.let {
-                imageCapture = it
-                useCasesList.add(it)
-            }
-        }
-
-        preview = plan.preview.also {
-            useCasesList.add(it)
-            it.surfaceProvider = environment.previewSurfaceProvider
+        if (isVideoMode) {
+            effects.setMicMutedIconVisible(!includeAudio)
         }
 
         effects.forceUpdateOrientationSensor()
 
-        // Not every camera can run video, photo and preview at once. Ask before binding rather
-        // than binding and retrying without the photo use case when it throws: an
-        // IllegalArgumentException from bindToLifecycle() carries no indication of which
-        // constraint was violated, so the retry could not tell "this camera can't do video plus
-        // photo" apart from any other misconfiguration, and would answer both by silently
-        // dropping in-video snapshots. A genuine bug then looked like a missing feature. This
-        // asks the specific question, and leaves unexpected exceptions to surface as failures.
-        //
-        // Asking costs 80-140 ms, because the camera service resolves the whole feature group to
-        // answer it, so the verdict is cached (snapshotSupport) and every entry into video mode
-        // after the first is free.
-        val snapshotUseCase = imageCapture
-        if (isVideoMode && snapshotUseCase != null) {
-            val probeKey = SnapshotProbeKey(
-                lensFacing = lensFacing,
-                videoQuality = videoQuality,
-                usesFeatureGroup = featureGroup is FeatureGroupRequest.Requested,
-                captureMode = plan.captureMode,
-                selectHighestResolution = selectHighestResolution,
+        val bindSettings = CameraBindSettings(
+            mode = currentMode,
+            isQrMode = isQRMode,
+            isVideoMode = isVideoMode,
+            requiresVideoModeOnly = entryPoint.requiresVideoModeOnly,
+            qrLensFacing = qrLensFacing,
+            rotation = rotation,
+            aspectRatio = aspectRatio,
+            flashMode = flashMode,
+            photoQuality = photoQuality,
+            videoQuality = videoQuality,
+            waitForFocusLock = waitForFocusLock,
+            enableZsl = enableZsl,
+            enableEis = enableEIS,
+            selectHighestResolution = selectHighestResolution,
+            mirrorVideoOnFrontCamera = saveVideoAsPreviewed,
+        )
+
+        return when (session.bind(bindSettings)) {
+            CameraSession.BindOutcome.FAILED -> effects.showMessage(R.string.bind_failure)
+
+            CameraSession.BindOutcome.EXTENSION_UNUSABLE -> {
+                effects.showMessage(R.string.extension_mode_unavailable)
+
+                // The bind never completed: currentMode still names the mode that was just
+                // disabled and nothing is rendering into the preview. Refreshing the tabs
+                // alone would only *visually* select another tab, leaving a frozen preview
+                // behind a lying tab bar. Switch for real -- switchMode() rebinds, moves the
+                // highlight and refreshes the tabs. Recursion stops because the default mode
+                // uses no extension.
+                switchMode(DEFAULT_CAMERA_MODE)
+            }
+
+            CameraSession.BindOutcome.BOUND -> announceBind()
+        }
+    }
+
+    private fun qrLensFacing(): Int {
+        if (session.isLensFacingSupported(
+                lensFacing = CameraSelector.LENS_FACING_BACK,
+                extensionMode = currentMode.extensionMode,
             )
-
-            if (!snapshotSupport.containsKey(probeKey)) {
-                snapshotProbeCount++
-
-                val cameraInfo = try {
-                    cameraProvider?.getCameraInfo(cameraSelector)
-                } catch (exception: IllegalArgumentException) {
-                    Log.e(TAG, "Failed to query camera info", exception)
-                    effects.showMessage(R.string.bind_failure)
-                    return
-                }
-
-                snapshotSupport[probeKey] = resolveInVideoSnapshotSupport(
-                    videoQualityFeature = requiredQualityFeature,
-                    probe = { withSnapshots, features ->
-                        val probedUseCases = when {
-                            withSnapshots -> useCasesList
-                            else -> useCasesList - snapshotUseCase
-                        }
-
-                        cameraInfo?.isSessionConfigSupported(
-                            SessionConfig(
-                                useCases = probedUseCases,
-                                requiredFeatureGroup = features,
-                            )
-                        ) == true
-                    },
-                )
-            }
-
-            val support = snapshotSupport[probeKey]
-            if (support is InVideoSnapshotSupport.Unsupported) {
-                Log.i(TAG, "${support.reason}; disabling snapshots while recording")
-                useCasesList.remove(snapshotUseCase)
-                imageCapture = null
-            }
+        ) {
+            return CameraSelector.LENS_FACING_BACK
         }
 
-        try {
-            val sessionConfig = SessionConfig(
-                useCases = useCasesList,
-                preferredFeatureGroup = plan.preferredFeatures
-            )
+        effects.showMessage(R.string.qr_rear_camera_unavailable)
 
-            if (plan.preferredFeatures.isNotEmpty()) {
-                val requested = plan.preferredFeatures.toList()
-                val boundLensFacing = lensFacing
-                sessionConfig.setFeatureSelectionListener(
-                    environment.sessionMainExecutor
-                ) { selected ->
-                    onFeaturesSelected(boundLensFacing, requested, requiredQualityFeature, selected)
-                }
-            }
+        return CameraSelector.LENS_FACING_FRONT
+    }
 
-            camera = cameraProvider!!.bindToLifecycle(
-                environment.sessionLifecycleOwner, cameraSelector,
-                sessionConfig
-            )
-        } catch (exception: RuntimeException) {
-            // A vendor extension can still fail to initialize at bind time even though the
-            // pre-flight probe in isExtensionUsable() passed. When one was applied, record the
-            // failure so the mode stops being offered, rather than letting the exception kill the
-            // process or -- equally bad -- retrying the same doomed bind on every resume.
-            val key = appliedExtension
-            if (key == null) {
-                // No extension in play: an IllegalArgumentException is a plain unsupported
-                // configuration (reported and swallowed); anything else is a real bug that must
-                // stay visible.
-                if (exception is IllegalArgumentException) {
-                    Log.e(TAG, "Failed to bind use cases", exception)
-                    effects.showMessage(R.string.bind_failure)
-                    return
-                }
-                throw exception
-            }
-
-            // With an extension applied, only the vendor's known-permanent signatures mean "this
-            // mode is unusable here": UnsupportedOperationException (a vendor extender that
-            // advertised the mode and then threw from its own init -- Pixels: "Framework size list
-            // map not supported in pixel path") and IllegalArgumentException (an extension bind
-            // always uses the same fixed pair of use cases, so an invalid configuration is as
-            // permanent as any other vendor failure). Anything else is a real bug and is rethrown
-            // rather than hidden behind a silent mode switch.
-            if (exception !is UnsupportedOperationException && exception !is IllegalArgumentException) {
-                throw exception
-            }
-
-            Log.e(TAG, "Extension mode $extMode failed to bind; disabling it", exception)
-            extensionAvailabilityStore.record(key, usable = false)
-            effects.showMessage(R.string.extension_mode_unavailable)
-
-            // The bind never completed: currentMode still names the mode that was just disabled
-            // and nothing is rendering into the preview. Refreshing the tabs alone would only
-            // *visually* select another tab, leaving a frozen preview behind a lying tab bar.
-            // Switch for real -- switchMode() rebinds, moves the highlight and refreshes the tabs.
-            // Recursion stops because the default mode uses no extension.
-            switchMode(DEFAULT_CAMERA_MODE)
-            return
-        }
-
+    private fun announceBind() {
         loadTabs()
 
-        // Every bind hands out a fresh LiveData in extension modes, and the old observer would
-        // otherwise stay attached: after a handful of mode switches a single zoom step redrew the
-        // thumb once per bind that had ever happened.
-        zoomStateSource?.removeObserver(zoomStateObserver)
-        zoomStateSource = null
-        zoomState = null
-        // Reading the new one is what costs ~100 ms behind an extension, and nothing before the
-        // next message needs it: the bar below draws a freshly bound camera's 1.0x either way, and
-        // every other reader is a gesture.
-        handler.removeCallbacks(attachZoomState)
-        handler.post(attachZoomState)
+        session.reattachZoomState()
 
         effects.updateZoomThumb(false)
 
@@ -1260,24 +760,9 @@ class CamConfig @AssistedInject constructor(
     private fun availableModes(): Set<CameraMode> {
         return resolveAvailableModes(
             allowsQrScanning = entryPoint.allowsQrScanning,
-            extensionsAvailable = extensionsManager != null && cameraProvider != null,
+            extensionsAvailable = session.extensionsAvailable,
         )
     }
-
-    private fun unprobedExtensions(): List<ExtensionKey> {
-        if (extensionsManager == null || cameraProvider == null) return emptyList()
-
-        return extensionAvailabilityStore.unprobed()
-    }
-
-    private fun selectorFor(lensFacing: Int): CameraSelector {
-        return when (lensFacing) {
-            CameraSelector.LENS_FACING_FRONT -> FRONT_CAMERA_SELECTOR
-            else -> REAR_CAMERA_SELECTOR
-        }
-    }
-
-    private var extensionProbesInFlight = false
 
     private fun loadTabs() {
         if (!entryPoint.showsCameraModeTabs) {
@@ -1293,42 +778,7 @@ class CamConfig @AssistedInject constructor(
         // tab bar still appears exactly once, fully formed, at the same time it used to; until
         // then swipes and taps resolve to no tab and the app simply stays in the current mode.
         // Later refreshes find the cache warm and rebuild synchronously, exactly as before.
-        val pending = unprobedExtensions()
-        if (pending.isEmpty()) {
-            buildTabs()
-            return
-        }
-
-        if (extensionProbesInFlight) return
-        val provider = cameraProvider ?: return
-        val em = extensionsManager ?: return
-        extensionProbesInFlight = true
-
-        thread {
-            val verdicts = HashMap<ExtensionKey, Boolean?>()
-            for (key in pending) {
-                verdicts[key] = probeExtension(
-                    provider, em, selectorFor(key.lensFacing), key.extensionMode
-                )
-            }
-
-            environment.sessionMainExecutor.execute {
-                extensionProbesInFlight = false
-                if (!environment.isSessionActive) return@execute
-
-                if (probedCameraProvider !== provider) {
-                    // The camera stack was reinitialized while probing: these verdicts describe
-                    // vendor state that no longer exists (the same reasoning as the cache clear
-                    // in initializeCamera). Any refresh that ran for the new provider found this
-                    // round still in flight and skipped scheduling, so start over for it.
-                    loadTabs()
-                    return@execute
-                }
-
-                extensionAvailabilityStore.recordProbeRound(verdicts)
-                buildTabs()
-            }
-        }
+        session.probeUnknownExtensions(onRestart = ::loadTabs, onSettled = ::buildTabs)
     }
 
     private fun buildTabs() {
@@ -1414,7 +864,7 @@ class CamConfig @AssistedInject constructor(
             }
         }
 
-        qrAnalyzer?.refreshHints()
+        session.refreshQrHints()
     }
 
     fun onStorageLocationNotFound() {
